backend/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
mailchimp3==3.0.13
psycopg2==2.8.5
django-modeltranslation==0.14.4
